Revert "Fix #135 : catch the new responses from Selenium 2.42"
This reverts f22f548 and 5d11c36. As discussed in #145, this solution
does not work as it interferes with some endpoints that expect to be
able to return false by indicating "" as their return value.

These fixes would catch that "" and return a non-empty string, which
would in turn evaluate to true.

Conflicts:
        lib/Selenium/Remote/RemoteConnection.pm